- Turned social into it's own component - Added component back to home page and cleaned unused code - Added social to footer - Fixed an issue on footer nav of missing page link.
