+++ components/Footer.tsx
@@ -1,6 +1,7 @@
 import Link from 'next/link'
 
 import { ContainerInner, ContainerOuter } from '@/components/Container'
+import SocialHorizontal from '@/components/SocialHorizontal'
 
 function NavLink({
   href,
@@ -28,10 +29,12 @@ export function Footer() {
             <div className="flex flex-col items-center justify-between gap-6 sm:flex-row">
               <div className="flex flex-wrap justify-center gap-x-6 gap-y-1 text-sm font-medium text-zinc-800 dark:text-zinc-200">
                 <NavLink href="/about">About</NavLink>
+                <NavLink href="/articles">Articles</NavLink>
                 <NavLink href="/projects">Projects</NavLink>
                 {/* <NavLink href="/speaking">Speaking</NavLink> */}
                 <NavLink href="/uses">Uses</NavLink>
               </div>
+              <SocialHorizontal />
               <p className="text-sm text-zinc-400 dark:text-zinc-500">
                 &copy; {new Date().getFullYear()} Rich Kevan. All rights
                 reserved.
